+++ plugin-build/src/main/kotlin/io/sentry/android/gradle/sourcecontext/OutputPaths.kt
@@ -2,21 +2,18 @@ package io.sentry.android.gradle.sourcecontext
 
 import org.gradle.api.Project
 
-internal const val ROOT_DIR = "reports/sentry-source-bundle"
+internal const val ROOT_DIR = "intermediates/sentry"
 
-/*internal*/ class OutputPaths(
+class OutputPaths(
     private val project: Project,
     variantName: String
 ) {
-
     private fun file(path: String) = project.layout.buildDirectory.file(path)
     private fun dir(path: String) = project.layout.buildDirectory.dir(path)
 
     private val variantDirectory = "$ROOT_DIR/$variantName"
-    private val intermediatesDir = "${variantDirectory}/intermediates"
-    private val sentryIntermediatesDir = "${intermediatesDir}/sentry"
 
-    val bundleIdDir = dir("${sentryIntermediatesDir}/bundle-id")
-    val sourceDir = dir("${sentryIntermediatesDir}/source-to-bundle")
-    val bundleDir = dir("${sentryIntermediatesDir}/source-bundle")
+    val bundleIdDir = dir("${variantDirectory}/bundle-id")
+    val sourceDir = dir("${variantDirectory}/source-to-bundle")
+    val bundleDir = dir("${variantDirectory}/source-bundle")
 }